pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tinyidp"
version = "1.1.1"
description = "A lightweight, configurable Identity Provider for development and testing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Christian Del Monte" }
]
keywords = [
    "identity-provider",
    "idp",
    "oauth2",
    "oidc",
    "saml",
    "authentication",
    "testing",
    "development",
    "mock",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "Flask>=3.0.0",
    "Flask-Cors>=4.0.0",
    "Flask-Limiter>=3.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "lxml>=5.0.0",
    "signxml>=3.2.0",
    "PyYAML>=6.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
tinyidp = "tinyidp.__main__:main"
tinyidp-mcp = "tinyidp.mcp_server:main"

[project.urls]
Homepage = "https://github.com/cdelmonte-zg/tinyidp"
Documentation = "https://github.com/cdelmonte-zg/tinyidp#readme"
Repository = "https://github.com/cdelmonte-zg/tinyidp.git"
Issues = "https://github.com/cdelmonte-zg/tinyidp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/tinyidp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/config",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
